guidance/agentic-orchestration/gateway/tools/textractor_tool/requirements.txt
boto3
amazon-textract-textractor
Pillow
